+++ admin/menu.php
@@ -24,6 +24,15 @@ add_action('admin_menu', function() {
 		'callback'		=> 'ziggeo_a_s_page')					//function
 	);
 
+	//Events Editor
+	ziggeo_p_add_menu_entry(array(
+		'page_title'	=> 'Ziggeo Events editor',				//page title
+		'menu_title'	=> 'Events Editor',						//menu title
+		'capability'	=> 'manage_options',					//min capability to view
+		'slug'			=> 'ziggeo_editor_events',				//menu slug
+		'callback'		=> 'ziggeo_a_ee_page')					//function
+	);
+
 	// Get notifications count
 	$n_count = ziggeo_notifications_count();
 	$n_title_extra = '';